Removing Standard Forms components from Angular.
