+++ .NetCore/Contoso/Contoso.Web.Flow/ScreenSettings/Views/ViewType.cs
@@ -13,7 +13,6 @@ namespace Contoso.Web.Flow.ScreenSettings.Views
         Delete,
         Html,
         About,
-        InputForm,
         FlowComplete,
         Exception
     }
