+++ Angular/Contoso/src/app/stuctures/screens/i-directive-function-args.ts
@@ -1,31 +1,5 @@
 import { FormGroup, FormBuilder } from "@angular/forms";
-import { IInputForm, IDirective, IInputQuestion } from "./input-form/i-input-form";
-import { IEditFormSettings, formTypeEnum, IFormItemSetting, IFormGroupData, IFormGroupSettings, IGroupSettings } from "./edit/i-edit-form-settings";
-
-export interface IInputDirectiveFunctionArgs {
-    directive: IDirective;
-    formGroup: FormGroup; 
-    fieldBeingUpdated: string; 
-    newValue: any;
-    targetControlName: string; 
-    targetControlFieldSetting: IInputQuestion
-    result: boolean;
-    conditionalDirectives: { [key: string]: IDirective[] };
-    formType: formTypeEnum;
-    fb: FormBuilder;
-}
-
-export interface IHandleInputDirectiveArgs {
-    directive: IDirective;
-    formGroup: FormGroup; 
-    fieldBeingUpdated: string; 
-    newValue: any;
-    targetControlName: string; 
-    targetControlFieldSetting: IInputQuestion
-    conditionalDirectives: { [key: string]: IDirective[] };
-    formType: formTypeEnum;
-    fb: FormBuilder;
-}
+import { IEditFormSettings, IFormItemSetting, IFormGroupData, IFormGroupSettings, IGroupSettings, IDirective } from "./edit/i-edit-form-settings";
 
 export interface IEditDirectiveFunctionArgs {
     directive: IDirective;
@@ -37,7 +11,6 @@ export interface IEditDirectiveFunctionArgs {
     targetControlFieldSetting: IFormItemSetting;
     result: boolean;
     conditionalDirectives: { [key: string]: IDirective[] };
-    formType: formTypeEnum;
     fb: FormBuilder;
 }
 
@@ -50,6 +23,5 @@ export interface IHandleEditDirectiveArgs {
     targetControlName: string; 
     targetControlFieldSetting: IFormItemSetting;
     conditionalDirectives: { [key: string]: IDirective[] };
-    formType: formTypeEnum;
     fb: FormBuilder;
 }
