+++ Angular/Enrollment/src/app/screen-host/screen-host.component.html
@@ -6,7 +6,6 @@
     <div [@flyIn]="'in'" *ngSwitchCase="viewType.Detail"><app-generic-detail [settings]="screenSettings.settings" [commandButtons]="screenSettings.commandButtons"></app-generic-detail></div>
     <div [@flyIn]="'in'" *ngSwitchCase="viewType.Delete"><app-generic-delete [settings]="screenSettings.settings" [commandButtons]="screenSettings.commandButtons"></app-generic-delete></div>
     <div [@flyIn]="'in'" *ngSwitchCase="viewType.Html"><app-html-page [settings]="screenSettings.settings" [commandButtons]="screenSettings.commandButtons"></app-html-page></div>
-    <div [@flyIn]="'in'" *ngSwitchCase="viewType.InputForm"><app-input-form [settings]="screenSettings.settings" [commandButtons]="screenSettings.commandButtons"></app-input-form></div>
     <div [@flyIn]="'in'" *ngSwitchCase="viewType.FlowComplete">FlowComplete</div>
     <div [@flyIn]="'in'" *ngSwitchCase="viewType.Exception">
       <div class="alert alert-danger">
