+++ Angular/Enrollment/src/app/stuctures/screens/i-view-type.ts
@@ -4,7 +4,6 @@ export enum ViewTypeEnum {
     Create,
     Detail,
     Delete,
-    InputForm,
     Html,
     FlowComplete,
     Exception
